move HotRestartImpl to source/common/hot_restart
