+++ source/common/hot_restart/BUILD
@@ -4,9 +4,27 @@ load(
     "//bazel:envoy_build_system.bzl",
     "envoy_cc_library",
     "envoy_package",
+    "envoy_select_hot_restart",
 )
 
 envoy_package()
+
+envoy_cc_library(
+    name = "hot_restart_lib",
+    srcs = envoy_select_hot_restart(["hot_restart.cc"]),
+    hdrs = envoy_select_hot_restart(["hot_restart.h"]),
+    deps = [
+        "//include/envoy/event:dispatcher_interface",
+        "//include/envoy/event:file_event_interface",
+        "//include/envoy/server:hot_restart_interface",
+        "//include/envoy/server:instance_interface",
+        "//include/envoy/server:options_interface",
+        "//source/common/common:assert_lib",
+        "//source/common/common:utility_lib",
+        "//source/common/network:utility_lib",
+        "//source/common/stats:stats_lib",
+    ],
+)
 
 envoy_cc_library(
     name = "hot_restart_nop_lib",
