+++ source/common/hot_restart/hot_restart.h
