+++ source/exe/BUILD
@@ -5,7 +5,6 @@ load(
     "envoy_cc_binary",
     "envoy_cc_library",
     "envoy_package",
-    "envoy_select_hot_restart",
 )
 
 envoy_package()
@@ -58,8 +57,8 @@ envoy_cc_library(
     srcs = ["main.cc"],
     deps = [
         ":envoy_main_common_lib",
-        ":hot_restart_lib",
         "//source/server:options_lib",
+        "//source/common/hot_restart:hot_restart_lib",
         "//source/server/config/http:lightstep_lib",
         "//source/server/config/http:zipkin_lib",
     ] + select({
@@ -74,27 +73,10 @@ envoy_cc_library(
     hdrs = ["main_common.h"],
     deps = [
         ":envoy_common_lib",
-        ":hot_restart_lib",
         "//source/common/common:compiler_requirements_lib",
+        "//source/common/hot_restart:hot_restart_lib",
         "//source/common/hot_restart:hot_restart_nop_lib",
         "//source/server/config_validation:server_lib",
-    ],
-)
-
-envoy_cc_library(
-    name = "hot_restart_lib",
-    srcs = envoy_select_hot_restart(["hot_restart.cc"]),
-    hdrs = envoy_select_hot_restart(["hot_restart.h"]),
-    deps = [
-        "//include/envoy/event:dispatcher_interface",
-        "//include/envoy/event:file_event_interface",
-        "//include/envoy/server:hot_restart_interface",
-        "//include/envoy/server:instance_interface",
-        "//include/envoy/server:options_interface",
-        "//source/common/common:assert_lib",
-        "//source/common/common:utility_lib",
-        "//source/common/network:utility_lib",
-        "//source/common/stats:stats_lib",
     ],
 )
